restructure to use character instead of calculator class
